+++ dd-java-agent/iast/src/main/java/com/datadog/iast/IastModuleImpl.java
@@ -11,7 +11,8 @@ import datadog.trace.util.stacktrace.StackWalkerFactory;
 
 public class IastModuleImpl implements IastModule {
   public void onCipherAlgorithm(String algorithm) {
-    if (Config.get().getWeakCipherAlgorithms().contains(algorithm.toUpperCase())) {
+    if (null != algorithm
+        && Config.get().getWeakCipherAlgorithms().contains(algorithm.toUpperCase())) {
       // get StackTraceElement for the callee of MessageDigest
       StackTraceElement stackTraceElement =
           StackWalkerFactory.INSTANCE.walk(
@@ -31,7 +32,8 @@ public class IastModuleImpl implements IastModule {
   }
 
   public void onHashingAlgorithm(String algorithm) {
-    if (Config.get().getWeakHashingAlgorithms().contains(algorithm.toUpperCase())) {
+    if (null != algorithm
+        && Config.get().getWeakHashingAlgorithms().contains(algorithm.toUpperCase())) {
       // get StackTraceElement for the callee of MessageDigest
       StackTraceElement stackTraceElement =
           StackWalkerFactory.INSTANCE.walk(
