--- dd-java-agent/instrumentation/java-security/src/test/groovy/test/UnavailableCipherAlgorithmTest.groovy
@@ -1,22 +1,0 @@
-package test
-
-import datadog.trace.agent.test.AgentTestRunner
-
-import javax.crypto.Cipher
-import java.security.NoSuchAlgorithmException
-
-import static datadog.trace.agent.test.utils.TraceUtils.runUnderTrace
-
-class UnavailableCipherAlgorithmTest extends AgentTestRunner {
-
-  def "test instrumentation"() {
-
-    when:
-    runUnderTrace("WeakHashingRootSpan") {
-      Cipher.getInstance("SHA-XXX")
-    }
-
-    then:
-    thrown NoSuchAlgorithmException
-  }
-}
